# make check-pylint: Speed up pylint