...by passing all python files to a single pylint process, instead of
one at a time. This way pylint doesn't have to re-load every imported
file (recursively) from scratch every time.

On my laptop this cuts the runtime of `make check-pylint` by 30 seconds
(from 53 seconds to 23 seconds) even though I was using 4 cores before
and now only 1 core. On Travis this saves about 70 seconds (the gains
are larger as there is less concurrency available on Travis).

Curiously this caused pylint to flag a few new problems in the
auto-generated files in `tests/auto-selftest-example-test-pack/selftest`
so I've added those to the list of excluded files.